+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
             android:theme="@style/Theme.UnCrack"/>
 
         <activity
-            android:name=".presentation.introScreens.SplashScreen"
+            android:name=".presentation.intro.SplashScreen"
             android:exported="true"
             android:theme="@style/Theme.UnCrack">
             <intent-filter>
@@ -43,7 +43,7 @@
         </activity>
 
         <activity
-            android:name=".presentation.introScreens.OnboardingScreen"
+            android:name=".presentation.intro.OnboardingScreen"
             android:exported="true"
             android:theme="@style/Theme.UnCrack">
         </activity>
