+++ app/src/main/java/com/geekymusketeers/uncrack/navigation/Navigation.kt
@@ -20,14 +20,14 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.geekymusketeers.uncrack.R
-import com.geekymusketeers.uncrack.presentation.accountScreen.AccountScreen
+import com.geekymusketeers.uncrack.presentation.account.AccountScreen
 import com.geekymusketeers.uncrack.presentation.home.HomeScreen
 import com.geekymusketeers.uncrack.presentation.masterKey.ConfirmMasterKeyScreen
 import com.geekymusketeers.uncrack.presentation.masterKey.CreateMasterKeyScreen
 import com.geekymusketeers.uncrack.presentation.masterKey.UpdateMasterKey
-import com.geekymusketeers.uncrack.presentation.passwordScreen.PasswordScreen
-import com.geekymusketeers.uncrack.presentation.profileScreen.ProfileScreen
-import com.geekymusketeers.uncrack.presentation.shieldScreen.ShieldScreen
+import com.geekymusketeers.uncrack.presentation.password.PasswordScreen
+import com.geekymusketeers.uncrack.presentation.profile.ProfileScreen
+import com.geekymusketeers.uncrack.presentation.shield.ShieldScreen
 import com.geekymusketeers.uncrack.ui.theme.BackgroundLight
 import com.geekymusketeers.uncrack.ui.theme.DMSansFontFamily
 import com.geekymusketeers.uncrack.ui.theme.FadeIn
