+++ app/src/main/java/com/geekymusketeers/uncrack/presentation/password/PasswordScreen.kt
@@ -1,4 +1,4 @@
-package com.geekymusketeers.uncrack.presentation.passwordScreen
+package com.geekymusketeers.uncrack.presentation.password
 
 import androidx.compose.runtime.Composable
